Split Encoder from Decoder
We can refactor into 2 classes and extract a superclass
to hold all the encoder/decoder sklearn models
