+++ src/exceptions/NegativePredictionLengthException.py
@@ -1,0 +1,2 @@
+class NegativePredictionLengthException(Exception):
+    pass
