+++ test/test_one_hot_matrix_decoder.py
@@ -1,0 +1,59 @@
+from unittest import TestCase
+
+import numpy as np
+
+from exceptions.NonpositiveLengthException import NonpositiveLengthException
+from onehot.OneHotMatrix import OneHotMatrixDecoder
+
+
+class TestOneHotMatrixDecoder(TestCase):
+    def test_encode_bad_length(self):
+        try:
+            decoder = OneHotMatrixDecoder(0)
+        except NonpositiveLengthException as e:
+            pass
+        except Exception as e:
+            self.fail()
+
+    def test_decode_sequences(self):
+        self.fail()
+
+    def test_decode_char(self):
+        encoder = OneHotMatrixDecoder(1)
+        encoded_sequences = np.array([
+            [
+                [1, 0, 0, 0]
+            ],
+            [
+                [0, 0, 0, 1]
+            ]
+        ])
+        decoded_sequences = encoder.decode_sequences(encoded_sequences)
+        expected_sequences = [
+            ["A"],
+            ["T"]
+        ]
+        np.testing.assert_array_equal(decoded_sequences, expected_sequences)
+
+    def test_decode_sequences(self):
+        encoder = OneHotMatrixDecoder(4)
+        encoded_sequences = np.array([
+            [
+                [0, 1, 0, 0],
+                [1, 0, 0, 0],
+                [0, 0, 0, 1],
+                [0, 0, 1, 0]
+            ],
+            [
+                [1, 0, 0, 0],
+                [0, 0, 0, 1],
+                [0, 0, 1, 0],
+                [0, 1, 0, 0]
+            ]
+        ])
+        decoded_sequences = encoder.decode_sequences(encoded_sequences)
+        expected_sequences = [
+            ["C", "A", "T", "G"],
+            ["A", "T", "G", "C"]
+        ]
+        np.testing.assert_array_equal(decoded_sequences, expected_sequences)
